@@ -4,17 +4,17 @@
     "private": true,
     "dependencies": {
         "@react-google-maps/api": "^2.19.3",
-        "@tanstack/react-query": "^5.37.1",
-        "@tanstack/react-query-devtools": "^5.37.1",
-        "@testing-library/jest-dom": "^6.4.2",
-        "@testing-library/react": "^14.2.2",
+        "@tanstack/react-query": "^5.39.0",
+        "@tanstack/react-query-devtools": "^5.39.0",
+        "@testing-library/jest-dom": "^6.4.5",
+        "@testing-library/react": "^15.0.7",
         "@testing-library/user-event": "^14.5.2",
         "leaflet": "^1.9.4",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
         "react-leaflet": "^4.2.1",
-        "react-router-dom": "5.3.4",
-        "styled-components": "^6.1.8"
+        "react-router-dom": "^6.23.1",
+        "styled-components": "^6.1.11"
     },
     "scripts": {
         "start": "vite",
@@ -37,7 +37,7 @@
         ]
     },
     "devDependencies": {
-        "@vitejs/plugin-react": "^4.2.1",
+        "@vitejs/plugin-react": "^4.3.0",
         "vite": "^5.2.11"
     }
 }
